First cut at the new AdaCore Style Checker (ASC)
It should support most kinds of files, except those we no longer
need to check (KP, features), and Java (to be implemented separately,
and hopefully soon).

This includes a testsuite, with coverage measurement. Coverage is
not complete yet (95%, or 24 lines still not covered), but the aim
is to have full coverage.

For NA17-007.