main.py: slight rework prepping for abort if too many files with errors
We want this style_checker to be able to accept multiple files
on the command-line so that we can call this checker once
for the entire list of files, rather than calling it once
for each and every file (performance).

THis is already supported with the current version. However, what
the current version does is just looping over each and every
file, and checks them all. This can create a lot of output when
checking in a change which modifies a lot of files with style
violations. So what we want instead is a way to stop the checking
after a certain threshold of files with style violations has been
reached.

This patch is just preparation towards that goal.

For PB01-008.