--dump-checks: Remove more useless columns.
'STYLE' is useless.

'FEATURES' is a typific that is no longe relevant (since move to
impactdb).

'START_BANG' is inherited from cvs_check and was useful to have
during the transition because it allowed direct textual diff between
the table cvs_check was dumping, and the one we did. But there was
no typific that used that check, as a matter of fact, so this was
never enabled, and thus we never implemented it in this style checker.
Now that the transition is over, just remove this column.