minor tweak to make copyright matching regexp configurable per repository
This is preparation work for making the regular expression used to
match copyright headers a configuration element. The tricky part
is that the regular expression is used to extract parts of it, based
on groups in the regular expression. At the moment, the groups are
identified by number, which will not be something we will be able
to rely on once the regexp becomes part of the repositories
configuration. So, instead, what this patch does is that it introduces
named groups that the code can refer to in order to extract the parts
it needs. Repositories that provide their own regexp will then have
to provide them in such a way that they use the same named groups:
  - 'holder' for the group that matches the name of the copyright holder
  - 'year' for the group matching the end year of the copyright year range.

For Q914-054.