Allow repositories to specify how copyright headers should be formatted
This patch turns some of the constants in rulific/copyright.py into
configuration entries to be ready from the style_checker's config file.

Note that it introduces one line of code currently not covered by
the testsuite. But this will come later, when we start updating
the configuration file with QGEN's copyright header info.

For Q914-054.