make the copyright configuration regexps list-based
This patch further enhanced the configuration abilities of the copyright
rulific checker by making the various entries list of strings or
list of regexps. The purpose is to allow the user to provide multiple
alternatives without having to hardcode that into one single regexp,
which could be quite tedious and result into an even less readable
regular expression.

For Q914-054.