Config::__read_config_file minor rewrite (no functional change)
This patch rewrites a bit the code in that function to factorize
access to the "system" configuration section (ANY_MODULE_NAME)
as well as access to the module-specific configuration section.
While this already makes the code slightly more readable, the real
purpose of that patch is to prepare for the introduction of
the module configuration file.

Part of QB13-061.