Remove erb-specific configuration
Since this repository is no longer active, no need to keep that
configuration in the style_checker repository.

This allows us to get rid of a ??? command in ada.py.

The testcase is not kept, as the features tested by this testcase
(ie - non-default copyright holders) are already exercised by other
tests.

For QC19-004.