Perform case-insensitive matching of the output of the "file" command.
This is necessary in the case of python scripts which do not have
a .py extension, because the output of the "file" command is different
depending on how the script's shebang is formatted. Two cases:

Case #1:

    $ head -1 a
    #! /usr/bin/env python
    $ file a
    a: a /usr/bin/env python script, ASCII text executable

Case #2:

    $ head -1 b
    #!/usr/bin/env python
    $ file b
    b: Python script, ASCII text executable

As we can see, the casing of "Python" changes between the two.
This patch fixes the issue by doing case-insensitive matching
of the output of the "file" command.

Change-Id: I48f46c807b85cb0cbff442b9473f01ec8b662c26