Explicit handling of .md files as markdown files (no checking for those)
A user reported that the style checker returned an error checking
the contents of a .md file, which is a markdown file. This shouldn't
happen, since we do not do any style-checking of markdown files.
Unfortunately, this did happen because we use the "file" command to
recognize scripts for which their filename has no telling extension,
and in this particular case, "file"'s heuristics make it return
that the file is a Python script (wrong). E.g:

    $ file 2_ENVIRONMENT.md
    2_ENVIRONMENT.md: Python script, ASCII text executable

Change-Id: Id2dab5a58bc36dfb30ba6b94af895ce45d3ddc9d
TN: RA22-066