Allow Ada 2005 constructs in gnat core compiler sources
This commit changes the default language version when checking
the Ada units which are part of the gnat core compiler sources.

A test file (bindo.adb) and the corresponding test file has been
added to verify that this is now allowed.

Another test is also added to check support for the 'gnat95'
config option. It becomes apparent that we're missing this test
when we change the default for core compiler Ada files to be
Ada 2005, because the coverage analysis then shows one less line
being covered.

Change-Id: Ic2cf3d59b47aac356e4298f6efb8cabaacccce9b
TN: T212-047