Make the "python_fragment" test work with newer versions of pycodestyle
Newer versions are stricter in terms of the placement of binary
operators, and now generate a warning when it is placed at the end
of a line. Older versions allowed both. This change fixes the various
test files to work with both.

This change also fixes the expected output for the error warning
when using an "except:" statement without the name of the exception(s)
to handle. Newer versions of pycodestyle were fixed to include a missing
single quote.

Change-Id: I58e93c0bb14b8ab8d846cb1d18ab9ce02970e366