TestCase.run_style_checker: be more specific about what **kwargs can be
Currently, this method accepts a **run_kwargs argument, with
documentation saying that these arguments are simply passed down
to the Run class. This gets in the way when one wants to add
more named arguments taking a default value which are not
related to the Run class, something I'd like to do in a separate
commit.

So, this commit first renames the **run_kwargs argument into
a **kwargs argument, to avoid the confusion that kwargs is
exclusively used in the "call" to the "Run" class.

It then modifies the implementation to iterate over the kwargs
dictionary, picking the arguments this method support, and raising
an exception for the arguments that it does not support. At the moment,
the only argument the testsuite uses is "input", so that's the only
one that's include in this commit. More will be needed if/when needed.

A note explaining why we rely on *args and **kwargs for this method's
signature is also added, to clarify the intent.

Change-Id: I362883b1f2481b8c43b68d38fd3b49feee6565ef
TN: T605-002