TestCase.run_style_checker: Enhance to allow calling via Python interpreter
This commit allows the few testcases that were calling the style_checker
directly to now do so by calling self.run_style_checker instead.
This centralizes the way the style_checker is called, which is
a useful enhancement in and of itself. The other upside is that
it should make it easier to transition the testsuite to a different
framework, one that is independent of gnatpython or e3, should we
want to (the framework itself is currently self-sufficient; it would
be nice if we could make the testsuite not depend on those either,
since this is a project used by contributors outside of AdaCore).

Change-Id: I3640c504426d23d363db16069551249946400e8d
TN: T605-002