fix python3 incompatibility in main.py
Fix one instance of a call to log_error where we pass an exception,
instead of passing the exception's args parameter (the function expects
either a string, or an iterable or strings).

This commit changes the call so as to be compatible with both Python 2.7
and Python 3.x.

Change-Id: I26c86d0c0a4d1c5dfb268bd9b0615be32a17bb33
TN: T605-002