config.py: fix Python3 compability issue iterating over dict keys
The Config.__read_config_file method iterates over a dictionary's keys
while at the same time modifying the dictionary itself. With Python 2,
the dictionary's "keys" method was returning a copy of these keys,
so this was no problem. However, with Python 3, this is no a copy.
As a result, when run with a Python 3 interpreter, this part of
the style_checker crashes with:

    RuntimeError: dictionary changed size during iteration

This commits makes the code compatible with both Python 2 and
Python 3.

Change-Id: I2a357707bbb12818e51388383ee5a4472aa61577
TN: T605-002