asclib/ex.py: Fix Python3 str/byte incompatilibity
With Python 3, the output buffer returned by Popen.communicate is
a bytes object, whereas all users of this class expect a string.
This change makes the code compatible for both Python 2 and Python 3,
by checking the type of the string, and coverting it to a string
assuming a UTF-8 encoding if not.

Note that this introduces one new line of code. The current testsuite,
which can only be run using gnatpython, which is Python-2-based,
does not cover that line. It will be covered when the testsuite is
converted over to pytest and made Python-3-based. In the meantime,
we temporarily accept that one line of code not being covered.

Change-Id: I476825b3c70f50cc9e8874bfd4f86d14f281885d
TN: T605-002