invalid DOS EOL not detected by style_checker when using Python 3
When running the style_checker through a Python 3 interpreter
(Python 3.8 to be precise), the style_checker fails to detect
invalid uses of DOS line endings.

This was detected by running the "dos_EOL" testcase from the current
testsuite, where the test_doseol_ko_1_adb test found the output
returned by the style_checker was missing the following errors:

    doseol-ko-1.adb:1: DOS line ending is not allowed
    doseol-ko-1.adb:2: DOS line ending is not allowed
    doseol-ko-1.adb:3: DOS line ending is not allowed [...]

This is caused by a change of default behavior in the "open" builtin
when reading text files, where line endings are now automatically
translated to '\n', instead of being left untouched.

This commit fixes the issue in a way that keeps the code compatible
with both Python 2.7 and Python 3.x.

Change-Id: I7a30c728455b78b7abb9ea7b2efc93ede35aab19
TN: T605-002 (umbreall TN for the transition to Python 3)
TN: T728-038 (TN for this specific compatibility issue)