Transition the testsuite to pytest and declare this tool as Python 3
This commit removes the old testsuite infrastructure, based on
a testsuite framework provided by gnatpython, and replaces it by
a pytest-based testsuite. This version of the testsuite requires
Python 3 to run, and proves that this tool is now Python 3-ready.
As we are not touching the tool's code in this commit (only
the testsuite), the tool remains compatible with Python 2 in practice,
but there is no intention in the future to preserve that compatibility.

The meat of the old infrastructure is now implemented via a new
fixture called "style_checker", which roughly provides the same
API as the old API in class TestCase. This way, the conversion
of each testcase from the old framework to the new one basically
consists in the following:

  - Renaming the test.py script into test_<name_of_testcase>.py
    (pytest requires that testscase scripts be named test_[...].py
    or [...]_test.py, and that testcase scripts all have a unique
    basename)

  - Stripping the creation of a TestCase class created by each
    testcase to contain all the methods implementing all the test
    of this testcase...

  - ... and moving these classes' methods to the root of the module
    as simple functions instead.

  - Textual search-and-replace "self" by "style_checker".

Outside of a few localized spots (related to verifying during
unit testing that an exception is being raised), everything
remained pretty much the same.

The expected output of some commands had to be modified to follow
the new output of more recent versions of Python, or associated
tools (better error messages, mostly).

A script called testsuite/run-validation-tests is provided to
make it simpler to run the testsuite as well as run all aspects
of the testing (including coverage and style checking).

Change-Id: Ie1f0849749ddb3b147de493931bba06882b2c3ea
TN: T605-002
(no-precommit-check because of one Python-3 feature tripping
the Python2-based style_checker currently deployed on our servers)