run-validation-tests: Add pycodestyle and pyflakes to list of requirements
These are requirements that this script checks for, ahead of running
the various parts of the testsuite. This avoids having testcases
reporting spurious failures which are due to the environment rather
than due to a bug in the style_checker. The two requirements being
added are not new, they were just missed when the list was first
created (I suspect we're missing more of them, but we'll update
the list as we discover them).

Found while working on T605-002 (transition to Python 3.x).

(no-precommit-check, because f-string elsewhere in script triggers
an error with the Python2-based style_checker used by our hooks)
ZZ
Change-Id: I4e495fde7d3187a1be9baa7bc6091bc8220d849a