run-validation-tests: Stop calling coverage report with -i
This was a workaround which appears to no longer be necessary.
Not entirely sure what it was that "fixed" the issue that the -i switch
was working around (whether the switch to using "python3 -m" to run
coverage, or maybe something related to how I managed both the old
pre-python3 testsuite framework combined with the new one).

(no-precommit-check, because f-string elsewhere in script triggers
an error with the Python2-based style_checker used by our hooks)

Change-Id: I68e44f7b7b21485fb36bf406ef6d6d9bbc2c7d1f
TN: T605-002