Call pycodestyle and pyflakes through sys.executable -m
With Python 3.x, using "python -m" is the recommended way to call these.
Beside this, another reason we are making this change is that this allows
us to make sure that we're calling the version of those tools that comes
with the Python interpreter that was used to execute the style_checker.
In particular, if we have both Python 2.x and Python 3.x in the PATH,
it prevents us from using the Python 2.x version of those tools. This is
something that will be particularly relevant during the transition period
where the style_checker is Python 3.x while some other tools run
at the same time (e.g. the git-hooks) are still Python 2.x.

This change does have an effect on a couple of testcases which were
testing the behavior of the style_check when either of these two tools
were present. The testing was done by manipulating the PATH to make it
effectively empty. This patch defeats this purpose since the style_checker
now uses sys.executable to run the tools it needs, and thus no longer
needs the PATH to be set. So a new testing strategy, centered around
creating a local Python distrib where pieces can be removed by hand,
is used in its place. See the new make_minimal_copy_of_python_install
method in class StyleCheckerFixture.

This did allow us to realize that it is no longer necessary to check
for OSError exceptions, since sys.executable cannot normally be missing.
So this allows a bit of code simplification as well.

(no-precommit-check, because f-string introduced in StyleCheckerFixture
triggers an error with the Python2-based style_checker used by our hooks)

Change-Id: Ied615b143346d700b1bb82c7f0285a14b1532b5e
TN: T605-002