Run 2to3 on all python fragment files used during testing
The main goal was to fix the code using print as a statment rather
than as a function. But while at it, it seemed worthwhile letting
2to3 do its magic. All changes looked good.

Found while working on T910-036, where I was experimenting with
the idea of using flake8 instead of pycodestyle only, which lead
to flak8 choking on the print statements (triggering a fatal
error -- E999 SyntaxError: invalid syntax).

Change-Id: I85144bb517d188afe701aa0b82382d369708beb9