Python: Explicitly exclude W504 errors from pycodestyle checks
Newer versions of pycodestyle has an ignore list which includes
both W503 (line break before binary operator) and W504 ((line break
after binary operator). So, by default, neither of these checks
are applied.

However, the style_checks is not using the default ignore list
(because we want to add E402 -- module level import not at top of file).
As a result, this causes W504 to get activated, which is not what
we wanted.

This change updates the ignore list to keep it in sync with
the default ignore list for recent versions of pycodestyle.

Change-Id: I9ff814250b1f4401a1f35c05b582824d9e9b973a
TN: TA05-042