Minor adjustments in a couple of python fragment test files
This commit makes a minor adjustment to a couple of test files used
by the "python_fragment" testcase, so as to avoid an F841 error
when running flake8 on them (local variable 'e' is assigned to but never
used").

This is preparation work for a future enhancement where the style
checker will switch to flake8 as the thirdparty checker.

Part of T910-036.

Change-Id: I77381e65191a8e13c8ab2ed289ac75e46791a5ed