transition the python checker from pycodestyle+pyflakes to flake8
This is a minor enhancement of the Python checker in the style_checker
where we now use flake8 instead of pycodestyle + flake8.

One small complication is the handling of what we call "Python
fragments" (see readme, and original TN for that feature: Q620-021).
If we wanted to be strictly equivalent to what we used to do before,
we would have to tell flake8 to ignore all the checks that pyflakes
would normally do. However, upon reviewing the various possible error
codes, I found that there was only one error that we obviously needed
to exclude. For the others (e.g. F823, which checks for undefined
*local* variables -- which therefore cannot come from the outside
context), it seems like those might actually be useful to keep,
even for fragments. This is therefore the approach taken by this change.

One small benefit from switching to flake8 is that the errors now
also include the error code when flagging an error.

One other small benefit of this change is that pycodestyle and
pyflakes checks are performed at the same time, so both are always
performed, meaning that errors from both are shown when running
the style_checker. This is better than what was before, because
the checker would stop if the first tool displayed some errors.
So a user might find himself in a situation where the checker
reports a few errors, leading the user to fix those thinking that
he'll be OK afterwards, only to find the style_checker reporting
*new* errors on the second attempt (because the first tool would
now pass, allowing the second to run and therefore find more errors).
We can see that in some tests, where the style_checker reports
more errors than before (all new errors were checked to make sure
they are indeed legitimate).

One final note: The pep8_missing and pyflakes_missing are now replaced
by one single test called pyflakes_missing, which is a straight copy
of those two testcases, where "pep8" or "pyflakes" is simple replaced
by "flake8".

Change-Id: Icb9415899fb345107048e17a50bee14cfaf3a9ec
TN: T910-036