run-validation-tests: Use pip3 to check for installed packages
Up until this commit, this script was testing for package installs
by checking whether the corresponding python module could be imported.
This has worked well enough so far, but a future commit will introduce
the requirement to have pytest-xdist also installed. This is a pytest
pluggin which enhances its capabilities, and it's not clear whether
it has a dedicated and well defined set of modules we can count on
to check whether it has been installed or not.

So this commit changes the way the run-validation-tests script
checks for dependencies, using pip3 instead.

For that, we first add pip3 to the list programs that this testsuite
needs. And we also move the check for missing programs first, so as
to allow the checking for missing packages to know whether it can
call pip3 or not.

To match the order in which the dependencies are checked,
the list of required packages is now moved after the list of
required programs. We're also updating the name of the modules
with the corresponding names of their packages.

Change-Id: I80634f90fd898b8f228726973863e97eadc377de