run-validation-tests: Run the style validation testing first...
... and abort the testing if this testing returned some errors.
Basically: We know that we're going to have to fix any style
violation found, which means modifying the code, which in turn
means we'll need to re-run the testsuite anyway. Might as well
just abort the testing early... Users who still want to do
the testing regardless of style violations can always bypass
this via --no-style-checking.

Change-Id: I2c68420243f36222a7f5a95102a20ca725b18d15