Rename all testcase scripts to all use the same name "run_test.py"
Up until recently, having unique filenames for the testcases was
required by pytest due to the way pytest was called and the testsuite
architected. But now that we're using --import-mode=importlib,
there is no such limitation anymore. So this commit renames all
current testcase scripts to a simpler, shorter and more consistent
"run_test.py".

For the record, the shell command used to do the renaming was:

    for i in */test_*.py; do
      tc_dir=`echo $i | cut -d/ -f1`
      git mv $i ${tc_dir}/run_test.py
    done

Change-Id: If8592e7812a2d8549d8c35a2272e93725f21377d