testsuite/conftest.py::env_setup: run testcase inside <tmp>/src
This commit fixes a small thinko: The "env_setup" fixture creates
a temporary directory, and then creates inside that temporary
directory two more subdirectories: A "tmp" directory, and an "src"
directory. The intent for that "src" directory was to be the working
directory inside which the testcase's sources would be copied and
the testcase itself be run.

The problem is that we os.chdir-ed to the temporary directory,
rather than the "src" subdirectory inside that temporary directory.
As a result, the testcase ended up not using our "src" subdirectory
at all. It's cleaner to run our testcase in that "src" subdirectory
because it is guaranteed to be empty at the time we copy the testcase's
sources over, so this commit fixes this issue by chdir-ing to
the "src" subdirectory.

Change-Id: I4047d805a7c658faee0cb8bc4f5094ae5110d61e