Add new "BIDI" rule to reject potentially dangerous unicode characters
A new rule is added to reject certain unicode characters that
an attacker could use to exploit the Unicode Bidi algorithm
and construct source code that would appear innocuous when rendered
in an editor, but that would actually contain malicious code.

This new rule is applied to all files which are known by
the style_checker to be a programming language.

Change-Id: Ic0e4a7d17d316effcfe9c74ef5b57b8bcf5b1779
TN: UB03-053