Force -gnatX for all GNAT runtime files
For Ada runtimes, we were relying on -gnatg to imply -gnatX
to allow GNAT extensions in those files. However, this isn't
always the case, so this commit changes the style_checker to
always call the compiler with -gnatX when checking a file
that's identified as a GNAT runtime file.

Change-Id: I913417e4b70b4cd13925b6f3b95887562a678bee
TN: UB09-015