Update various tests' expected output after V103-009
In V103-009, I updated the list of valid copyright holders
to add a couple more names, but clearly I forgot to run
the testsuite, because this of course causes some changes
to the style_checker's output when a file should have a valid
copyright, but does not.

This commit fixes that omission, allowing us to get clean testing
again.

Change-Id: I652d88d81cb91c04126cb0fc00219ab1fb1a85e0
TN: V103-009