Run the testsuite with -vv to get more detailed info in case of error
This commit adds -vv to the list of command-line arguments we pass
to pytest when running the testsuite. This doesn't change anything
when the test is successful, but increases the amount of information
being provided when a test fails. In particular, when a string equality
test fails, it'll show the whole diff, rather than just an abbreviated
version of the expanded strings. Without this, if the strings are
are a few lines long, it becomes impossible to diagnose the issue
just from the testsuite report alone.

Change-Id: I5ecdbb3a9badf12f33ea9fd1091f6e170310492c