run-validation-tests: disable PIP version check when calling pip3
The run-validation-tests has a pre-testing phase where it verifies
that all necessary dependencies for running the testsuite without
error are present. One of those checks is the list of Python modules
being installed, which we get by calling pip3, asking for the list
in JSON format.

This commit avoids a testsuite crash when the version of pip being
installed is not the latest one:

| $ ./run-validation-tests
| Traceback (most recent call last):
| [...]
|   File "/[...]/./run-validation-tests", line 117, in check_dependencies
|     package_list_from_pip3 = json.loads(p.out)
| [...]
|     raise JSONDecodeError("Extra data", s, end)
| json.decoder.JSONDecodeError: Extra data: line 3 column 1 (char 5379)

The error happens because pip added the following warning
after the JSON output:

    | $ pip3 list --format=json
    | [JSON data snipped]
    |
    | [notice] A new release of pip available: xxxx -> yyyy
    | [notice] To update, run: python -m pip install --upgrade pip

This commit avoids this issue by telling pip3 to not do a version check.

Found while working on W105-011.

Change-Id: I00cfd6203d113c28e6ddb2991d6eae3834308f8b